pyproject.toml
[tool.poetry]
name = "TTL"
version = "1.1.0"
description = "Simple TODO Task List App"
authors = ["Ismael Mousa <ismaelramzimousa@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = { extras = ["standard"], version = "^0.25.0" }
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
testfixtures = "^7.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "*.py"
norecursedirs = "__init__.py"


[tool.coverage.run]
source = ["."]
omit = ["main.py", "*/__init__.py", "tests/*", ".venv/*", "routers/task.py"]


[tool.pylint]
disable = ["C0114", "C0209", "W0511", "W0707", "W0718", "W1203", "W1514", "R0801", "R0903", "R0904"]
ignore-paths = [".venv"]
max-line-length = 100
